Extract assistant advisory facade helpers
